Add `V1Collector`
