+++ src/collectors/sketchCollector.ts
@@ -1,4 +1,5 @@
-import { AbstractCollector, Invoice, CompleteInvoice, CollectorType, CollectorState, CollectorCaptcha } from "./abstractCollector";
+import { Invoice, CompleteInvoice, CollectorType, CollectorState, CollectorCaptcha } from "./abstractCollector";
+import { V1Collector } from "./v1Collector";
 import { UnfinishedCollectorError } from '../error';
 import { Location } from "../proxy/abstractProxy";
 import { Secret } from "../secret_manager/abstractSecretManager";
@@ -22,7 +23,7 @@ export type SketchConfig = {
     }
 }
 
-export abstract class SketchCollector extends AbstractCollector {
+export abstract class SketchCollector extends V1Collector {
 
     constructor(config: SketchConfig) {
         super({
